Update unit tests to work with or without Subsites
As the rest of the module also follows this ethos. Adapted the CI tools to
install all the integration dependencies (otherwise optional) for
completeness. Also updated a number of class calls missed in initial
upgrade, generally because of being contained as a string (e.g. in a
`class_exists` call).